third_party/WebKit/LayoutTests/fast/multicol/balance-line-underflow-1.html

<!DOCTYPE html>
<p>There should be a blue square below.</p>
<div id="multicol" data-expected-height="20" style="columns:2; width:20px; line-height:20px; background:blue;">
    <div style="line-height:60px; margin-top:-40px;"><br></div>
    <br>
</div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<script>
    checkLayout("#multicol");
</script>
